move log restoration to kvmap and parsing to operation
