+++ src/kvmap/map_error.rs
@@ -3,6 +3,7 @@ pub enum MapError {
     KeyNotFound(String),
     InvalidKey(String),
     EmptyValue(String),
+    RestoreError(String),
 }
 
 impl MapError {
@@ -11,6 +12,7 @@ impl MapError {
             MapError::KeyNotFound(key) => format!("Key not found: {}", key),
             MapError::InvalidKey(key) => format!("Invalid key: {}", key),
             MapError::EmptyValue(key) => format!("No value found at: {}", key),
+            MapError::RestoreError(error) => format!("Failed to restore db from file: {}", error),
         }
     }
 }
