+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod kvmap;
 pub mod operation;
-pub mod parser;
 pub mod radixtree;
 pub mod wal;
