--- src/operation.rs
@@ -1,6 +1,0 @@
-#[derive(Debug, PartialEq)]
-pub enum Operation {
-    Get(String),
-    Put(String, String),
-    Delete(String),
-}
